rename list_sessions file to sessions
